WMS/pyproject.toml
[project]
name = "WMS"
version = "0.1.0"
description = "A WMS project for learning"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # --- Core Framework ---
    "fastapi>=0.128.0",
    "uvicorn>=0.40.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.8.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.9",
    # --- Database ---
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.20.0",
    "psycopg2-binary>=2.9.9",
    # --- Testing & Client ---
    "httpx>=0.28.1",
    "requests>=2.32.5",
    "pytest>=9.0.2",
    "locust>=2.43.1",
    "email-validator>=2.3.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.hatch.build.targets.editable]
packages = ["app"]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
